+++ go.mod
@@ -3,6 +3,7 @@ module github.com/zema1/suo5
 go 1.20
 
 require (
+	github.com/chainreactors/proxyclient v1.0.1-0.20250214091837-3dc4d4a0a010
 	github.com/go-gost/gosocks5 v0.3.0
 	github.com/gobwas/glob v0.2.3
 	github.com/kataras/golog v0.1.8
@@ -11,14 +12,13 @@ require (
 	github.com/refraction-networking/utls v1.6.4
 	github.com/stretchr/testify v1.10.0
 	github.com/urfave/cli/v2 v2.27.4
-	github.com/zema1/rawhttp v0.1.1
+	github.com/zema1/rawhttp v0.2.0
 )
 
 require (
 	github.com/andybalholm/brotli v1.1.0 // indirect
 	github.com/chainreactors/files v0.0.0-20231102192550-a652458cee26 // indirect
 	github.com/chainreactors/logs v0.0.0-20241115105204-6132e39f5261 // indirect
-	github.com/chainreactors/proxyclient v1.0.1-0.20250213161237-aabcbe2aeb47 // indirect
 	github.com/cloudflare/circl v1.3.8 // indirect
 	github.com/cpuguy83/go-md2man/v2 v2.0.4 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
@@ -31,8 +31,5 @@ require (
 	github.com/xrash/smetrics v0.0.0-20240521201337-686a1a2994c1 // indirect
 	golang.org/x/crypto v0.24.0 // indirect
 	golang.org/x/sys v0.21.0 // indirect
-	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
-
-replace github.com/zema1/rawhttp => github.com/M09Ic/rawhttp v0.1.2
