+++ main.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/zema1/suo5/suo5"
 	"os"
 	"os/signal"
 	"strings"
@@ -22,7 +23,7 @@ func main() {
 	app.Usage = "A high-performance http tunnel"
 	app.Version = Version
 
-	defaultConfig := ctrl.DefaultSuo5Config()
+	defaultConfig := suo5.DefaultSuo5Config()
 	app.DisableSliceFlagSeparator = true
 
 	app.Flags = []cli.Flag{
@@ -157,7 +158,7 @@ func Action(c *cli.Context) error {
 	target := c.String("target")
 	noAuth := c.Bool("no-auth")
 	auth := c.String("auth")
-	mode := ctrl.ConnectionType(c.String("mode"))
+	mode := suo5.ConnectionType(c.String("mode"))
 	ua := c.String("ua")
 	bufSize := c.Int("buf-size")
 	timeout := c.Int("timeout")
@@ -178,7 +179,7 @@ func Action(c *cli.Context) error {
 	if auth == "" {
 		if !noAuth {
 			username = "suo5"
-			password = ctrl.RandString(8)
+			password = suo5.RandString(8)
 		}
 	} else {
 		parts := strings.Split(auth, ":")
@@ -189,7 +190,7 @@ func Action(c *cli.Context) error {
 		password = parts[1]
 		noAuth = false
 	}
-	if !(mode == ctrl.AutoDuplex || mode == ctrl.FullDuplex || mode == ctrl.HalfDuplex) {
+	if !(mode == suo5.AutoDuplex || mode == suo5.FullDuplex || mode == suo5.HalfDuplex) {
 		return fmt.Errorf("invalid mode, expected auto or full or half")
 	}
 
@@ -212,7 +213,7 @@ func Action(c *cli.Context) error {
 		}
 	}
 
-	config := &ctrl.Suo5Config{
+	config := &suo5.Suo5Config{
 		Listen:           listen,
 		Target:           target,
 		NoAuth:           noAuth,
